perf(PianoRoll): delete self-managed pixel caches, converge to JUCE single-path direct painting
Root cause: GenerationSignature included ppsMilli/pixelsPerSemitone, causing tiles_.clear() + 48-128MB sync rebuild per wheel event.

Architecture: Deleted all self-managed pixel caches (TimelineCompositeCache, compositeCache_, viewportSurface_, pianoKeySurface_, activeOverlayItem_, 17 zoom deferral fields, 16 functions). paint() directly draws all 10 layers per frame. Constructor enables setBufferedToImage(true). Ctrl/Shift zoom: zero Image allocation, zero tile operations.

Additional: worldTopY int->float, unvoiced bands visible-frame-only scan, unused variable cleanup.

Tests: All pass.

+++ Source/Standalone/UI/TimelineLayerComposer.cpp
@@ -268,7 +268,7 @@ void TimelineLayerComposer::drawLaneStripRepeats(juce::Graphics& g, const Render
     const auto themeId = static_cast<ThemeId>(params.themeId);
     const float pixelsPerSemitone = params.pixelsPerSemitone;
     const int pianoKeyWidth = 0;
-    const int worldTopY = params.worldTopY;
+    const float worldTopY = params.worldTopY;
     const bool showLanes = decodeShowLanes(params.laneStyle);
     const int scaleRootNote = decodeScaleRootNote(params.laneStyle);
     const int scaleType = decodeScaleType(params.laneStyle);
